pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-faf-mcp"
version = "1.0.1"
authors = [
  { name="wolfejam", email="james@wolfejam.dev" },
]
description = "Universal Context Landing Pad for FAF (Foundational AI-context Format)"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["faf", "mcp", "gemini", "ai-context", "model-context-protocol"]
dependencies = [
    "functions-framework>=3.0.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-secret-manager>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
]

[project.urls]
"Homepage" = "https://github.com/Wolfe-Jam/gemini-faf-mcp"
"Glory Wall" = "https://faf-landing.vercel.app/glory.html"
"FAF Specification" = "https://faf.one"
"IANA Registration" = "https://www.iana.org/assignments/media-types/application/vnd.faf+yaml"
"Bug Tracker" = "https://github.com/Wolfe-Jam/gemini-faf-mcp/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json"]
